+++ packages/rnative/src/utils/helper.ts
@@ -1,3 +1,5 @@
+import URL from 'url-parse';
+
 export const rem = (strings: TemplateStringsArray): number => {
   const rootFontSize = 16;
   const remValue = parseFloat(strings[0]); // Extract the value from the template literal