+++ packages/rnative/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @lezzform/react-native
 
+## 0.0.13
+
+### Patch Changes
+
+- add external url-parse
+
+## 0.0.12
+
+### Patch Changes
+
+- log uploaded data
+
+## 0.0.11
+
+### Patch Changes
+
+- try using formData
+
+## 0.0.10
+
+### Patch Changes
+
+- experimental network error fix
+
+## 0.0.9
+
+### Patch Changes
+
+- change docpicker to file object
+
+## 0.0.8
+
+### Patch Changes
+
+- convert document picker response to file object
+
+## 0.0.7
+
+### Patch Changes
+
+- testing error attachment
+
+## 0.0.6
+
+### Patch Changes
+
+- attachment experimental native lib
+
+## 0.0.5
+
+### Patch Changes
+
+- attachment element init
+
 ## 0.0.4
 
 ### Patch Changes